exchange products

+++ ROMS/Model/TRN_ReturnDC.cs
@@ -31,13 +31,13 @@ namespace ROMS.Model
         public int ParaSubGroupID { get; set; } = 0;
         public int paraReasonId { get; set; } = 0;
         public int paraClosingReasonId { get; set; } = 0;
-        public float paraReturnDCAmount { get; set; } = 0;
+        public double paraReturnDCAmount { get; set; } = 0;
         public string paraCreditNoteNo { get; set; } = "";
         public string paraCreditNoteDate { get; set; } = "";
         public string paraRemarks { get; set; } = "";
-        public float paraDeleteFlag { get; set; } = 0;
-        public float ParaSubtotal { get; set; } = 0;
-        public float paraTax { get; set; } = 0;
+        public double paraDeleteFlag { get; set; } = 0;
+        public  double ParaSubtotal { get; set; } = 0;
+        public double paraTax { get; set; } = 0;
         public DataTable paraTRN_Purchase_ReturnDC { get; set; } = null;
     }
 }
